icecap-qemu: Fixed time issue due to a missing init call
Turns out veracruz_runtime_init was an important function.

Though it seems to only be needed for reading NOW and setting up
the current time, which is needed for certificate expiration. This
raises some questions about if this is fit for purpose. One
alternative would be for vc-server to send over the current time
during realm creation.